revert new template ctor, not needed
