+++ src/base/perf.hpp
@@ -25,7 +25,6 @@
 
 #include "perf_def.hpp"
 
-#include <stdexcept>
 #include <string>
 #include <type_traits>
 
@@ -114,17 +113,8 @@ namespace cubperf
 	COUNTER_AND_TIMER
       };
 
-      // constructor
       inline constexpr stat_definition (const stat_id id, type stat_type,
-					const char *first_name, const char *second_name = nullptr)
-	: m_id (id)
-	, m_type (stat_type)
-	, m_names { first_name, second_name }
-	, m_offset (0)
-      {
-      }
-
-      // copy constructor
+					const char *first_name, const char *second_name = nullptr);
       stat_definition (const stat_definition &other);
 
       stat_definition &operator= (const stat_definition &other);
@@ -149,7 +139,7 @@ namespace cubperf
   //
   // see how to use in file description comment
   //
-  class statset_definition
+  class statset_definition final
   {
     public:
       using stat_definition_init_list_t = std::initializer_list<stat_definition>;
@@ -158,10 +148,6 @@ namespace cubperf
       // no default constructor
       statset_definition (void) = delete;
       statset_definition (stat_definition_init_list_t defs);
-
-    private:
-      template <typename TI>
-      statset_definition (std::size_t a_size, TI a_begin, TI a_end);
 
     public:
       statset_definition (const statset_definition &) = delete;
@@ -270,46 +256,13 @@ namespace cubperf
   // statset_definition
   //////////////////////////////////////////////////////////////////////////
 
-  template <typename TI>
-  statset_definition::statset_definition (std::size_t a_size, TI a_begin, TI a_end)
-    : m_stat_count (a_size)
-    , m_value_count (0)
-    , m_stat_defs (nullptr)
-    , m_value_names (nullptr)
+  inline constexpr stat_definition::stat_definition (const stat_id id, type stat_type,
+      const char *first_name, const char *second_name)
+    : m_id (id)
+    , m_type (stat_type)
+    , m_names { first_name, second_name }
+    , m_offset (0)
   {
-    // copy definitions
-    m_stat_defs = new stat_definition[a_size];
-    std::size_t stat_index = 0;
-    for (TI it = a_begin; it != a_end; ++it)
-      {
-	auto &def_it = *it;
-	if (def_it.m_id != stat_index)
-	  {
-	    // statset_definition is bad; crash program
-	    throw std::runtime_error ("statset_definition is bad");
-	  }
-	m_stat_defs[stat_index] = def_it;  // copy definitions
-
-	// set offset and increment value count
-	m_stat_defs[stat_index].m_offset = m_value_count;
-	m_value_count += def_it.get_value_count ();
-
-	// increment index
-	stat_index++;
-      }
-
-    // names for all values
-    m_value_names = new std::string[m_value_count];
-    std::size_t value_index = 0;
-    for (stat_index = 0; stat_index < m_stat_count; stat_index++)
-      {
-	assert (value_index == m_stat_defs[stat_index].m_offset);
-	for (std::size_t def_name_index = 0; def_name_index < m_stat_defs[stat_index].get_value_count ();
-	     def_name_index++)
-	  {
-	    m_value_names[value_index++] = m_stat_defs[stat_index].m_names[def_name_index];
-	  }
-      }
   }
 
   template <bool IsAtomic>
